Move parallel runner code into seperate files
