+++ src/harness/tsconfig.json
@@ -92,6 +92,9 @@
         "loggedIO.ts",
         "rwcRunner.ts",
         "test262Runner.ts",
+        "./parallel/shared.ts",
+        "./parallel/host.ts",
+        "./parallel/worker.ts",
         "runner.ts",
         "../server/protocol.ts",
         "../server/session.ts",
